requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.23.0
gunicorn>=21.2.0
pydantic
pydantic[email]
pydantic-settings
email-validator
sqlalchemy>=2.0
alembic
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]
httpx
google-generativeai
groq
openai
python-multipart
python-dotenv
requests
python-docx
openpyxl
pandas
pypdf
pillow
tabulate
schedule
resend
waitress
langdetect
